+++ src/element.jl
@@ -1,0 +1,28 @@
+
+abstract type AbstractElement end
+
+
+mutable struct StreamElement <: AbstractElement
+end
+
+
+mutable struct StreamRecord <: AbstractElement
+    value
+    timestamp::Int
+    hasTimestamp::Bool
+end
+
+
+function isWatermark(record::StreamRecord)::Bool
+    
+end
+
+function getValue(record::StreamRecord)
+    return record.value
+end
+
+
+
+
+
+